주석 추가 및 익스텐션 분리
